@@ -6,7 +6,7 @@ version: 0.10.1+2
 publish_to: none
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -24,7 +24,7 @@ dependencies:
   path_provider: ^2.0.9
   file_picker: ^4.0.0
   file_selector: ^0.8.4+1
-  autocomplete_textfield: ^1.7.3
+  autocomplete_textfield: ^2.0.1
   csv: ^5.0.0
   archive: ^3.1.2
   windows_documents: ^1.0.0
