@@ -3,13 +3,13 @@ import 'package:flutter/services.dart';
 
 class AttributeTable extends StatelessWidget {
   final Map<String, String> attributes;
-  final double dataRowHeight;
+  final double? dataRowHeight;
 
   const AttributeTable({
-    Key key,
+    super.key,
     this.attributes = const <String, String>{},
     this.dataRowHeight,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -23,9 +23,10 @@ class AttributeTable extends StatelessWidget {
             .map<DataRow>((String key) => DataRow(
                   onSelectChanged: (_) => _onTapRow(context, key),
                   cells: [
-                    DataCell(Text(key, style: TextStyle(fontWeight: FontWeight.bold))),
+                    DataCell(Text(key,
+                        style: TextStyle(fontWeight: FontWeight.bold))),
                     DataCell(
-                      Text(attributes[key], overflow: TextOverflow.clip),
+                      Text(attributes[key] ?? '', overflow: TextOverflow.clip),
                     )
                   ],
                 ))
@@ -35,8 +36,9 @@ class AttributeTable extends StatelessWidget {
   }
 
   Future<void> _onTapRow(BuildContext context, String key) async {
-    await Clipboard.setData(ClipboardData(text: attributes[key]));
-    ScaffoldMessenger.of(context).hideCurrentSnackBar(reason: SnackBarClosedReason.hide);
+    await Clipboard.setData(ClipboardData(text: attributes[key] ?? ''));
+    ScaffoldMessenger.of(context)
+        .hideCurrentSnackBar(reason: SnackBarClosedReason.hide);
     ScaffoldMessenger.of(context).showSnackBar(SnackBar(
       content: Text('$key wurde in die Zwischenablage kopiert'),
       duration: const Duration(seconds: 1),
