@@ -1,26 +1,26 @@
 import 'package:flutter/material.dart';
 
 class GesturelessListTile extends StatelessWidget {
-  final Widget leading;
-  final Widget title;
-  final Widget trailing;
-  final Widget subtitle;
+  final Widget? leading;
+  final Widget? title;
+  final Widget? trailing;
+  final Widget? subtitle;
 
   /// padding for the whole content of the tile
-  final EdgeInsetsGeometry contentPadding;
+  final EdgeInsetsGeometry? contentPadding;
 
   /// makes setting abound height for the tile possible
-  final double height;
+  final double? height;
 
   const GesturelessListTile({
-    Key key,
+    super.key,
     this.leading,
     this.title,
     this.trailing,
     this.subtitle,
     this.contentPadding,
     this.height,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -32,7 +32,7 @@ class GesturelessListTile extends StatelessWidget {
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         crossAxisAlignment: CrossAxisAlignment.center,
         children: <Widget>[
-          leading,
+          leading ?? const SizedBox(),
           Expanded(
             child: Column(
               mainAxisAlignment: MainAxisAlignment.spaceEvenly,
@@ -42,22 +42,22 @@ class GesturelessListTile extends StatelessWidget {
                 if (title != null)
                   RichText(
                     text: WidgetSpan(
-                      style: Theme.of(context).textTheme.headline6,
-                      child: title,
+                      style: Theme.of(context).textTheme.headlineSmall,
+                      child: title!,
                     ),
                   ),
                 if (subtitle != null)
                   RichText(
                     text: WidgetSpan(
-                      style: Theme.of(context).textTheme.subtitle1,
-                      child: subtitle,
+                      style: Theme.of(context).textTheme.bodyMedium,
+                      child: subtitle!,
                     ),
                   ),
-              ].where((Widget w) => w != null).toList(),
+              ],
             ),
           ),
-          trailing,
-        ].where((Widget w) => w != null).toList(),
+          trailing ?? const SizedBox(),
+        ],
       ),
     );
   }
