@@ -9,26 +9,36 @@ import 'base_shortcut.dart';
 class ItemShortcut extends StatelessWidget {
   final BuildContext context;
   final Item item;
-  final Vendor vendor;
+  final Vendor? vendor;
   final bool showVendor;
 
-  const ItemShortcut(this.context,
-      {Key key, @required this.item, this.vendor, this.showVendor = true})
-      : super(key: key);
+  const ItemShortcut(
+    this.context, {
+    super.key,
+    required this.item,
+    this.vendor,
+    this.showVendor = true,
+  });
 
   @override
   Widget build(BuildContext context) {
     return BaseShortcut(
       context,
       onTap: () => Navigator.push<bool>(
-          context, MaterialPageRoute(builder: (BuildContext context) => ItemPage(item: item))),
+          context,
+          MaterialPageRoute(
+              builder: (BuildContext context) => ItemPage(item: item))),
       children: <Widget>[
         Text(item.title,
-            style: Theme.of(context).textTheme.subtitle1.copyWith(fontWeight: FontWeight.bold),
+            style: Theme.of(context)
+                .textTheme
+                .bodyMedium!
+                .copyWith(fontWeight: FontWeight.bold),
             textScaleFactor: 1.1,
             overflow: TextOverflow.ellipsis),
-        Text(item.description ?? '', overflow: TextOverflow.ellipsis),
-        if (showVendor && vendor != null) Text(vendor.name, overflow: TextOverflow.ellipsis),
+        Text(item.description, overflow: TextOverflow.ellipsis),
+        if (showVendor && vendor != null)
+          Text(vendor!.name, overflow: TextOverflow.ellipsis),
         Text(formatFigure(item.price), overflow: TextOverflow.ellipsis),
       ],
     );
