@@ -2,12 +2,14 @@
 //  Generated file. Do not edit.
 //
 
+// clang-format off
+
 #include "generated_plugin_registrant.h"
 
-#include <file_chooser/file_chooser_plugin.h>
+#include <sentry_flutter/sentry_flutter_plugin.h>
 
 void fl_register_plugins(FlPluginRegistry* registry) {
-  g_autoptr(FlPluginRegistrar) file_chooser_registrar =
-      fl_plugin_registry_get_registrar_for_plugin(registry, "FileChooserPlugin");
-  file_chooser_plugin_register_with_registrar(file_chooser_registrar);
+  g_autoptr(FlPluginRegistrar) sentry_flutter_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "SentryFlutterPlugin");
+  sentry_flutter_plugin_register_with_registrar(sentry_flutter_registrar);
 }
