@@ -7,9 +7,9 @@ import Foundation
 
 import file_selector_macos
 import package_info_plus_macos
-import path_provider_macos
+import path_provider_foundation
 import sentry_flutter
-import sqflite
+import sqflite_darwin
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   FileSelectorPlugin.register(with: registry.registrar(forPlugin: "FileSelectorPlugin"))
